pyproject.toml
[project]
name = "diagram-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.24.1",
    "python-dotenv>=1.2.1",
]
